+++ extensions/virtfusion/src/types.ts
@@ -8,21 +8,21 @@ export type Panel = {
 export type Server = {
   id: string;
   name: string;
-  bootOrder: Array<"hd"|"cdrom">
+  bootOrder: Array<"hd" | "cdrom">;
   memory: string;
   cpu: string;
   storage: Array<{
     capacity: string;
-    "enabled": boolean
-    "primary": boolean
+    enabled: boolean;
+    primary: boolean;
   }>;
   network: {
     primary: {
       limit: string;
       ipv4: Array<{
         address: string;
-        gateway: string
-        netmask: string
+        gateway: string;
+        netmask: string;
       }>;
       ipv6: Array<{
         addresses: string[];
@@ -41,14 +41,14 @@ export type SSHKey = {
 };
 
 export type Task = {
-  "action": string
-  "started": string
-  "updated": string
-  "finished": string
-  "completed": boolean,
-  "status": string
-  "success": boolean
-}
+  action: string;
+  started: string;
+  updated: string;
+  finished: string;
+  completed: boolean;
+  status: string;
+  success: boolean;
+};
 
 export type SingleResult<T> = {
   data: T;