+++ extensions/virtfusion/src/virtfusion.ts
@@ -16,7 +16,7 @@ class VirtFusion {
       ...options,
     });
     if (!response.ok) {
-      const result = await response.json().catch(() => null) as {errors: string[]} | null
+      const result = (await response.json().catch(() => null)) as { errors: string[] } | null;
       throw new Error(result?.errors[0] || response.statusText);
     }
     if (response.headers.get("Content-Type")?.includes("text/html") || response.status === 204) return undefined as T;
@@ -27,23 +27,23 @@ class VirtFusion {
   public async connect() {
     return this.request<void>("connect");
   }
-  public async setBootOrder(params: { serverId: string}, payload: {order: string }) {
+  public async setBootOrder(params: { serverId: string }, payload: { order: string }) {
     return this.request<{
-  "data": {
-    "task": {
-      "id": number,
-      "status": string
-    }
-  }
-}>(`server/${params.serverId}/bootOrder`, {
-  method: "POST",
-  body: JSON.stringify(payload)
-});
+      data: {
+        task: {
+          id: number;
+          status: string;
+        };
+      };
+    }>(`server/${params.serverId}/bootOrder`, {
+      method: "POST",
+      body: JSON.stringify(payload),
+    });
   }
   public async listServers(params: { page: number }) {
     return this.request<PaginatedResult<Server>>(`server?page=${params.page}`);
   }
-  public async listServerTasks(params: { serverId: string, page: number }) {
+  public async listServerTasks(params: { serverId: string; page: number }) {
     return this.request<PaginatedResult<Task>>(`server/${params.serverId}/tasks/?page=${params.page}`);
   }
   public async listAccountSSHKeys(params: { page: number }) {