+++ extensions/virtfusion/CHANGELOG.md
@@ -1,3 +1,10 @@
 # VirtFusion Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Manage Panels
+    1. List Servers
+        - Server Details
+            - Flip Boot Order
+    2. List SSH Keys
+        - Delete SSH Key